[chandra] rendered views dynamically inside adapter

+++ app/src/main/java/com/example/tmobile/model/TMobileHomePages.kt
@@ -44,7 +44,7 @@ data class Description(
 
 @Parcelize
 data class Font(
-    val size: Int
+    val size: Float
 ) : Parcelable, TMobileBusinessObject()
 
 @Parcelize
